+++ banzai/crosstalk.py
@@ -11,10 +11,6 @@ class CrosstalkCorrector(Stage):
     def __init__(self, pipeline_context):
         super(CrosstalkCorrector, self).__init__(pipeline_context)
 
-    @property
-    def group_by_attributes(self):
-        return None
-
     def do_stage(self, images):
         for image in images:
             if image.data_is_3d():