+++ banzai/photometry.py
@@ -21,10 +21,6 @@ class SourceDetector(Stage):
     def __init__(self, pipeline_context):
         super(SourceDetector, self).__init__(pipeline_context)
 
-    @property
-    def group_by_attributes(self):
-        return None
-
     def do_stage(self, images):
         for i, image in enumerate(images):
             try: