+++ banzai/tests/test_dark_maker.py
@@ -1,12 +1,10 @@
-from __future__ import absolute_import, division, print_function, unicode_literals
-from banzai.utils import stats
-from banzai.dark import DarkMaker
+import mock
 import numpy as np
 from astropy.io import fits
 
+from banzai.utils import stats
+from banzai.dark import DarkMaker
 from banzai.tests.utils import FakeImage, FakeContext, throws_inhomogeneous_set_exception
-
-import mock
 
 
 class FakeDarkImage(FakeImage):
@@ -23,7 +21,7 @@ def test_min_images():
     assert len(processed_images) == 0
 
 
-def test_group_by_keywords():
+def test_group_by_attributes():
     maker = DarkMaker(None)
     assert maker.group_by_attributes == ['ccdsum']