+++ banzai/tests/test_header_checker.py
@@ -23,11 +23,6 @@ def test_no_input_images():
     assert len(images) == 0
 
 
-def test_group_by_keywords():
-    tester = header_checker.HeaderSanity(None)
-    assert tester.group_by_attributes is None
-
-
 def test_all_keywords_missing():
     logger.error = mock.MagicMock()
     tester = header_checker.HeaderSanity(None)