+++ banzai/stages.py
@@ -1,6 +1,6 @@
 import logging
 import abc
-
+from banzai import logs
 import elasticsearch
 
 from banzai.utils.qc import format_qc_results
@@ -21,8 +21,8 @@ class Stage(abc.ABC):
         logger.info('Running {0}'.format(self.stage_name), image=images[0])
         try:
             images = self.do_stage(images)
-        except Exception as e:
-            logger.error(e)
+        except Exception:
+            logger.error(logs.format_exception())
         return images
 
     @abc.abstractmethod
@@ -54,7 +54,8 @@ class Stage(abc.ABC):
                                       doc_type=self.pipeline_context.elasticsearch_doc_type,
                                       id=filename, body={'doc': results_to_save, 'doc_as_upsert': True},
                                       retry_on_conflict=5, timestamp=results_to_save['@timestamp'], **kwargs)
-            except Exception as e:
+            except Exception:
                 error_message = 'Cannot update elasticsearch index to URL \"{url}\": {exception}'
-                logger.error(error_message.format(url=self.pipeline_context.elasticsearch_url, exception=e))
+                logger.error(error_message.format(url=self.pipeline_context.elasticsearch_url,
+                                                  exception=logs.format_exception()))
         return es_output