+++ CHANGES.md
@@ -1,9 +1,13 @@
+0.14.1 (2018-11-13)
+-------------------
+- Added full traceback of uncaught exceptions to the logs
+- Removed group_by_attributes property fom all stages except CalibrationMaker
+- Added master_selection_criteria property to CalibrationComparer
+  
 0.14.0 (2018-11-13)
 -------------------
 - Refactored bias, dark, and flat makers to use a common superclass to remove
   code duplication.
-- Removed group_by_attributes property fom all stages except CalibratrionMaker
-- Added master_selection_criteria property to CalibrationComprarer
 
 0.13.0 (2018-11-12)
 -------------------